feature/hw-1-edit-structure
