+++ homework1/cli.py
@@ -16,22 +16,17 @@ class TicTacToeGame:
     def show_board(self):
         self.board.show()
 
-    def validate_input(self, vertically, horizontally):
-        if (horizontally <= 0 or horizontally > self.board.size or vertically <= 0
-                or vertically > self.board.size or self.board.cells[vertically - 1][horizontally - 1] != '.'):
-            raise InvalidMoveException("Invalid move")
-
     def move(self, turn):
         print(self.MOVE_REQUEST)
         try:
             vertically, horizontally = map(int, input().split())
         except ValueError:
             raise InvalidMoveException("Move should be two integers")
-        self.validate_input(vertically, horizontally)
+        self.board.validate_input(vertically, horizontally)
         self.board.cells[vertically - 1][horizontally - 1] = turn
 
     def start(self, turn='x'):
-        while not self.check_draw():
+        while self.board.available_cells:
             self.show_board()
             self.move(turn)
             if self.check_winner(turn):
@@ -41,44 +36,13 @@ class TicTacToeGame:
             turn = '0' if turn == 'x' else 'x'
         print(self.CONGRATULATION_DRAW)
 
-    def check_horizontally(self, horizontally, turn):
-        for i in range(self.board.size):
-            if self.board.cells[i][horizontally] != turn:
-                return False
-        return True
-
     def check_winner(self, turn):
         for i in range(self.board.size):
             for j in range(self.board.size):
-                if (self.check_horizontally(j, turn) or self.check_vertically(i, turn) or
-                        self.check_diagonally(turn) or self.check_rev_diagonally(turn)):
+                if (self.board.check_horizontally(j, turn) or self.board.check_vertically(i, turn) or
+                        self.board.check_diagonally(turn) or self.board.check_rev_diagonally(turn)):
                     return True
         return False
-
-    def check_draw(self):
-        for i in self.board.cells:
-            for j in i:
-                if j == '.':
-                    return False
-        return True
-
-    def check_vertically(self, vertically, turn):
-        for i in range(self.board.size):
-            if self.board.cells[vertically][i] != turn:
-                return False
-        return True
-
-    def check_diagonally(self, turn):
-        for i in range(self.board.size):
-            if self.board.cells[i][i] != turn:
-                return False
-        return True
-
-    def check_rev_diagonally(self, turn):
-        for i in range(self.board.size):
-            if self.board.cells[self.board.size - 1 - i][i] != turn:
-                return False
-        return True
 
 
 if __name__ == '__main__':
